requirements.txt
alembic==1.15.2
authlib==1.5.1
Flask==3.1.0
flask_cors==5.0.1
flask_mail==0.10.0
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
protobuf==6.30.2
python-dotenv==1.1.0
SQLAlchemy==2.0.40
Werkzeug==3.1.3
google-generativeai==0.8.4
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
oauthlib==2.1.0